CONFIG: added ui extenions
