+++ HomeVue/Configs/UIColorExtension.swift
@@ -1,0 +1,16 @@
+//
+//  UIColorExtension.swift
+//  Home Vue
+//
+//  Created by student-2 on 09/12/24.
+//
+
+import Foundation
+import UIKit
+
+extension UIColor {
+    static let gradientStartColor = UIColor(red: 0x63/255.0, green: 0x56/255.0, blue: 0x55/255.0, alpha: 1.0) // Hex #302A29
+    static let gradientEndColor = UIColor(red: 0x54/255.0, green: 0x49/255.0, blue: 0x48/255.0, alpha: 1.0) // Hex #544948
+
+    static let solidBackgroundColor = UIColor(red: 0xD9/255.0, green: 0xD9/255.0, blue: 0xD9/255.0, alpha: 1.0) // Hex #D9D9D9
+}
